refactor: Legacy DB schema to declarative approach
