--- Setup/InstallSchema.php
@@ -1,148 +1,0 @@
-<?php
-/**
- * Taxjar_SalesTax
- *
- * NOTICE OF LICENSE
- *
- * This source file is subject to the Open Software License (OSL 3.0)
- * that is bundled with this package in the file LICENSE.txt.
- * It is also available through the world-wide-web at this URL:
- * http://opensource.org/licenses/osl-3.0.php
- *
- * @category   Taxjar
- * @package    Taxjar_SalesTax
- * @copyright  Copyright (c) 2017 TaxJar. TaxJar is a trademark of TPS Unlimited, Inc. (http://www.taxjar.com)
- * @license    http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
- */
-
-namespace Taxjar\SalesTax\Setup;
-
-use Magento\Framework\Setup\InstallSchemaInterface;
-use Magento\Framework\Setup\ModuleContextInterface;
-use Magento\Framework\Setup\SchemaSetupInterface;
-
-/**
- * @codeCoverageIgnore
- */
-class InstallSchema implements InstallSchemaInterface
-{
-    /**
-     * {@inheritdoc}
-     * @SuppressWarnings(PHPMD.ExcessiveMethodLength)
-     */
-    // @codingStandardsIgnoreStart
-    public function install(SchemaSetupInterface $setup, ModuleContextInterface $context)
-    {
-        // @codingStandardsIgnoreEnd
-        $installer = $setup;
-        $installer->startSetup();
-
-        /**
-         * Update table 'tax_class'
-         */
-        $installer->getConnection()->addColumn(
-            $installer->getTable('tax_class'),
-            'tj_salestax_code',
-            [
-                'type' => \Magento\Framework\DB\Ddl\Table::TYPE_TEXT,
-                'nullable' => false,
-                'default' => '',
-                'after' => 'class_type',
-                'comment' => 'Class Tax Code for TaxJar'
-            ]
-        );
-
-        /**
-         * Create table 'tax_nexus'
-         */
-        $table = $installer->getConnection()->newTable(
-            $installer->getTable('tax_nexus')
-        )->addColumn(
-            'id',
-            \Magento\Framework\DB\Ddl\Table::TYPE_INTEGER,
-            null,
-            ['identity' => true, 'unsigned' => true, 'nullable' => false, 'primary' => true],
-            'ID'
-        )->addColumn(
-            'api_id',
-            \Magento\Framework\DB\Ddl\Table::TYPE_INTEGER,
-            null,
-            ['unsigned' => true],
-            'API ID'
-        )->addColumn(
-            'street',
-            \Magento\Framework\DB\Ddl\Table::TYPE_TEXT,
-            255,
-            [],
-            'Street'
-        )->addColumn(
-            'city',
-            \Magento\Framework\DB\Ddl\Table::TYPE_TEXT,
-            255,
-            [],
-            'City'
-        )->addColumn(
-            'country_id',
-            \Magento\Framework\DB\Ddl\Table::TYPE_TEXT,
-            2,
-            [],
-            'Country Id'
-        )->addColumn(
-            'region',
-            \Magento\Framework\DB\Ddl\Table::TYPE_TEXT,
-            255,
-            [],
-            'Region'
-        )->addColumn(
-            'region_id',
-            \Magento\Framework\DB\Ddl\Table::TYPE_INTEGER,
-            null,
-            [],
-            'Region Id'
-        )->addColumn(
-            'region_code',
-            \Magento\Framework\DB\Ddl\Table::TYPE_TEXT,
-            255,
-            [],
-            'Region Code'
-        )->addColumn(
-            'postcode',
-            \Magento\Framework\DB\Ddl\Table::TYPE_TEXT,
-            255,
-            [],
-            'Postcode'
-        )->addColumn(
-            'created_at',
-            \Magento\Framework\DB\Ddl\Table::TYPE_TIMESTAMP,
-            null,
-            [
-                'nullable' => false,
-                'default' => \Magento\Framework\DB\Ddl\Table::TIMESTAMP_INIT
-            ],
-            'Created At'
-        )->addColumn(
-            'updated_at',
-            \Magento\Framework\DB\Ddl\Table::TYPE_TIMESTAMP,
-            null,
-            [
-                'nullable' => false,
-                'default' => \Magento\Framework\DB\Ddl\Table::TIMESTAMP_INIT_UPDATE
-            ],
-            'Updated At'
-        )->addIndex(
-            $installer->getIdxName('tax_nexus', ['country_id']),
-            ['country_id']
-        )->addIndex(
-            $installer->getIdxName('tax_nexus', ['region_id']),
-            ['region_id']
-        )->addIndex(
-            $installer->getIdxName('tax_nexus', ['region_code']),
-            ['region_code']
-        )->setComment(
-            'TaxJar Nexus Address'
-        );
-        $installer->getConnection()->createTable($table);
-
-        $installer->endSetup();
-    }
-}
